requirements.txt
altair==5.2.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autopep8==2.0.4
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-js-asset==2.1.0
django-mptt==0.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
executing==2.0.1
Faker==20.1.0
fastjsonschema==2.19.1
filelock==3.12.0
fonttools==4.48.1
fqdn==1.5.1
frozendict==2.4.0
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.1.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
multitasking==0.0.11
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
notebook_shim==0.2.3
numpy==1.26.2
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
peewee==3.17.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==3.5.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.21.12
psutil==5.9.8
psycopg==3.1.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycodestyle==2.11.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
